+++ src/main/java/de/samply/notification/smtp/EmailTemplateType.java
@@ -1,0 +1,6 @@
+package de.samply.notification.smtp;
+
+public enum EmailTemplateType {
+    INVITATION,
+    NEW_PROJECT
+}